refactor: no need for classes
