+++ conftest.py
@@ -1,20 +1,18 @@
 import pytest
 
-from tinystatic.steps.copy_static_assets_step import CopyStaticAssetsStep
-from tinystatic.steps.generate_pages_step import GeneratePagesStep
-from tinystatic.steps.prepare_environment_step import PrepareEnvironmentStep
+from tinystatic.steps import copy_static_assets, generate_pages, prepare_environment
 
 
 @pytest.fixture
 def prepare_env_step():
-    return PrepareEnvironmentStep()
+    return prepare_environment.run
 
 
 @pytest.fixture
 def generate_pages_step():
-    return GeneratePagesStep()
+    return generate_pages.run
 
 
 @pytest.fixture
 def copy_static_assets_step():
-    return CopyStaticAssetsStep()
+    return copy_static_assets.run
